Avoid calling `buildFinished` after build is finished
Search links for custom values must be added during a 'buildFinished' event,
in order to ensure that the server URL is set to it's final value.
The previous implementation did this by registering a separate 'buildFinished'
listener for each search link, but the background process that captured git metadata
would occasionally register after the build was complete.

In order to avoid the warnings issued in this case, we now register a single
buildFinished listener, with a thread-safe implementation that will collect all
registered build scans. With this change, if the git metadata process completes after
the build is complete, then the search link is ignored and no warning is issued.

Note that configuration-cache compliance necessitated the use of a static class
to register as a buildFinished listener.