Proxy Gradle Enterprise plugin types to work around ClassLoader issues
When the GE plugin is applied by the build and the CCUD plugin via an
init script, the Gradle Enterprise extension and types it references
are loaded by a different classloader than the classes in the CCUD
plugin. To work around this issue, the Gradle Enterprise types are now
accessed via dynamic proxies that "translate" between the different
class loaders at runtime.