Refactor CI environment variable checks.
The goal of this is to reduce existing code duplication. This will make it easier in the future to refactor all the CI environment variable checks to be done inside a static buildFinished context so that they are configuration cache compatible.
If the approach is okay, we will also need to do the same refactor for the CCUD Maven extension.