Match only if remote repo hostname contains github or gitlab
Previously, the remote repo regex matcher would match on any https or ssh repo url and use regex groups to parse out the hostname.
However, since we only create build scan links to GitHub or GitLab remote repos, there is no reason for the matcher to match at all if the remote repo is not a GitHub or GitLab url.