Re-implement extracting host and path from repository url using regex
URI was not able to parse the standard ssh urls provided by GitHub when cloning a repository.
These urls can be handled more easily with a regex.
Though harder to read, implementing the host and path extraction with regex allows for a wider range of urls to be parsed, and allows for creating custom links to more varieties of remote repository urls.