Support build scan links to alternate GitHub URLs (#150)
* Use GITHUB_SERVER_URL env var for CaptureCiMetadataAction.execute

* Dynamically create github/gitlab links using regex groups

* Update changelog

* Handle GITHUB_SERVER_URL not set in GitHub Actions

Defaults to https://github.com if running in GitHub Actions and GITHUB_SERVER_URL is unset

* Match only if remote repo hostname contains github or gitlab

Previously, the remote repo regex matcher would match on any https or ssh repo url and use regex groups to parse out the hostname.
However, since we only create build scan links to GitHub or GitLab remote repos, there is no reason for the matcher to match at all if the remote repo is not a GitHub or GitLab url.

* Update changes.md to describe fixed build scan links issue #109

* Make capitalization consistent for 'GitHub/GitLab source' links

* Move orElse(https://github.com) to gitHubUrl assignment

This way all logic for setting the value of gitHubUrl is contained in one location

* Do not match on http git remote urls

When creating build scan links to remote git repository, only match https repository urls

* Use GitHub Actions environment variables to generate remote source build scan link

If running in GitHub Actions, the build scan link to the remote github repository will be generated using GITHUB_SERVER_URL and GITHUB_REPOSITORY environment variables. This allows build scan links to enterprise GitHub installations or other GitHub installations with custom URLs - even those URLs that do not contain "github"

* Update matcher to also match on ssh:// urls and other ssh usernames

* Do not match on ssh://

* Use non-capturing regex groups

Use non-capturing regex groups for those parts of the regex we don't intend to capture

* Use regex to ignore .git extension in repo url

Use regex rather than using substring() to trim .git from the end of the remote repo url

* Update release notes to use correct tense

* Fix build scan link generator uses optional instead of contained value

* Narrow scope of if statement

Narrow scope of if statment to only encapsulate what is necessary

* Use URI instead of regex to generate "GitHub source" build scan link

URI will validate the remote repository url returned by `git config --get remote.origin.url`. Additionally, it is much clearer than using a regex with groups.

* Reduce scope of try/catch when creating URI of remote repo url

* Check hostname for github/gitlab when generating custom link to remote repo

hostname is the relevant scope of the search, not the entire origin url

* Remove fallback for github.com

* Extract method to convert String to URI

* Rename variables from `gitHubRepository` to `gitRepository`

* Implement toHttpsUri method to construct custom link

* Made conversion from remote git repository url to url for custom link more concise

* Coerce invalid ssh github URIs into valid URIs

* Fix invalid git@github.com and git@gitlab.com repository URIs

If the remote repository URI begins with git@github.com or git@gitlab.com, this is an invalid URI

* Re-implement extracting host and path from repository url using regex

URI was not able to parse the standard ssh urls provided by GitHub when cloning a repository.
These urls can be handled more easily with a regex.
Though harder to read, implementing the host and path extraction with regex allows for a wider range of urls to be parsed, and allows for creating custom links to more varieties of remote repository urls.

* Refactor slightly and non-functionally

* Improve readability

---------

Co-authored-by: Etienne Studer <etienne@gradle.com>